+++ Updater/source/__init__.py
@@ -5,3 +5,4 @@ from source import download
 from source import exceptions
 from source import creds
 from source import pmc
+from source import gui
