+++ Updater/source/gui/utils.py
